fix: satisfy pkg-file-count and backend-size CI limits
Fold agypty corpus loader into cleaner.go, merge cleaner tests into argv_test.go, and share Agy provider test stubs via inference_provider_test_helpers_test.go.
